Character service added

+++ Angular/angular-dndcustom/src/app/services/http.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpInterceptor, HttpRequest, HttpHandler, HttpParams, HttpResponse } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import * as _ from 'lodash';
 
 const httpOptions = {
   headers: new HttpHeaders({ 'Content-Type': 'application/json' })
@@ -10,32 +11,32 @@ export class HttpService {
 
   constructor(private httpClient: HttpClient) { }
 
-  save<T = any>(api: string, controller: string, method: string, data: any): Observable<T> {
+  save<T = any>(controller: string, method: string, data: any): Observable<T> {
     const body = JSON.stringify(data);
     return this.httpClient.post<T>(this.getActionUrl(controller, method), body, httpOptions);
   }
 
-  put<T = any>(api: string, controller: string, method: string, data: any) : Observable<T> {
+  put<T = any>(controller: string, method: string, data: any) : Observable<T> {
     const body = JSON.stringify(data);
     return this.httpClient.put<T>(this.getActionUrl(controller, method), body, httpOptions);
   } 
-  post<T = any>(api: string, controller: string, method: string, data: any, responseType: any = "json"): Observable<T> {
+  post<T = any>(controller: string, method: string, data: any, responseType: any = "json"): Observable<T> {
     const body = JSON.stringify(data);
     return this.httpClient.post<T>(this.getActionUrl(controller, method), body, { responseType, ...httpOptions });
   } 
-  postResponse<T = any>(api: string, controller: string, method: string, data: any, responseType: any = "json"): Observable<HttpResponse<T>> {
+  postResponse<T = any>(controller: string, method: string, data: any, responseType: any = "json"): Observable<HttpResponse<T>> {
     const body = JSON.stringify(data);
     return this.httpClient.post<T>(this.getActionUrl(controller, method), body, { responseType, observe: 'response', ...httpOptions });
   } 
-  delete<T>(api: string, controller: string, method: string, param: any): Observable<T> {
+  delete<T>(controller: string, method: string, param: any): Observable<T> {
     // the reason of this code is RESTApi requires request as following pattern api/controller/id
     // therefore, we have to put the id into generated url
-    /*if (_.isObject(param)) {
+    if (_.isObject(param)) {
       // when we are sending an object with DELETE still send it with other ways. 
       return this.httpClient.delete<T>(this.getActionUrl(controller, method), { params: new HttpParams({ fromObject: param })  });
     } else {
       return this.httpClient.delete<T>(this.getActionUrl(controller, method) + param);
-    }*/
+    }
     return;
   }
 
@@ -63,29 +64,9 @@ export class HttpService {
     return this.httpClient.get(url);
   }
 
-  getClaims(claims) {
-    return this.httpClient.get(this.getActionUrl('User', 'GetUserClaims?sessionId=' + sessionStorage.getItem('sessionRef') + '&claims=' + claims.join()));
-  }
-
-  getUrl(api: string, controller: string, method: string) {
+  getUrl(controller: string, method: string) {
     return this.getActionUrl(controller, method);
   }
 }
 
-@Injectable()
-export class TokenInterceptor implements HttpInterceptor {
 
-  constructor() { }
-
-  intercept(request: HttpRequest<any>, next: HttpHandler): any {
-    console.log(sessionStorage.getItem('currentUser'));
-    request = request.clone({
-      setHeaders: {
-        Authorization: 'Bearer ' + sessionStorage.getItem('currentUser')
-      }
-    });
-    // TODO: https://ryanchenkie.com/angular-authentication-using-the-http-client-and-http-interceptors #Looking for Unauthorized Responses
-    // Add auth control in here.
-    // REASON: CanActivate is only work when the routing happens. Since, we have ajax request in pages we can get 401 at this point, we shouldn't handle in components.
-  }
-}
